fix(rust_scorer): widen synapse from_index u16 to u32 for GPU mirror
neat-core's SynapseExport.from_index changed from u32 to u16, breaking
the rust_scorer release build (E0308: expected u32, found u16) at
forward_mse_batched.rs:228. The GPU/WGSL SynapseGpu mirror must stay u32
(WGSL has no 16-bit int), so widen with u32::from(...), matching the
adjacent num_synapses handling.

This unblocks NEAT-AI-Examples CI, whose unit-tests job builds rust_scorer
from NEAT-AI-scorer@Develop (stSoftwareAU/NEAT-AI-Examples#606).

Vibe-Coder-Run-Id: vibe-mqqg2mpj-c37c8f

+++ rust_scorer/src/gpu/forward_mse_batched.rs
@@ -225,7 +225,9 @@ pub fn build_batched_network_data(
         for s in &net.synapses {
             synapses.push(SynapseGpu {
                 weight: s.weight,
-                from_index: s.from_index,
+                // neat-core's SynapseExport.from_index is u16; the GPU/WGSL
+                // mirror is u32 (WGSL has no 16-bit int), so widen here.
+                from_index: u32::from(s.from_index),
             });
         }
